fixed simple bags in scss

+++ .eslintrc.js
@@ -15,7 +15,7 @@ module.exports = {
   plugins: ["react", "prettier"],
   rules: {
     "react/react-in-jsx-scope": "off",
-    quotes: ["error", "single", { allowTemplateLiterals: true }],
+    quotes: ["error", "double", { allowTemplateLiterals: true }],
     "jsx-quotes": ["error", "prefer-double"],
     "react/prop-types": "off",
     "no-unused-vars": "warn"
